+++ hfut/value.py
@@ -1,7 +1,8 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
+
 import re
-from functools import wraps
+
 import six
 
 HF = 'HF'